+++ backend/src/main/java/com/blog/blogspringboot/service/CommentService.java
@@ -2,13 +2,17 @@ package com.blog.blogspringboot.service;
 
 import com.blog.blogspringboot.dto.CommentRequestDTO;
 import com.blog.blogspringboot.entity.Comment;
+import com.blog.blogspringboot.entity.User;
 import com.blog.blogspringboot.repository.CommentRepository;
+import com.blog.blogspringboot.util.UserUtils;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
+import java.security.Principal;
 import java.util.Date;
 
 @Service
@@ -45,8 +49,19 @@ public class CommentService {
         return commentRepository.findAll(pageable);
     }
 
-    public void deleteComment(int id) {
-        commentRepository.deleteById(id);
+    public HttpStatus deleteComment(int id, Principal principal) {
+        User user = userService.getUserByUsername(principal.getName());
+        Comment comment = getCommentById(id);
+        if (comment == null) {
+            return HttpStatus.NOT_FOUND;
+        }
+        boolean isModerator = UserUtils.isUserModerator();
+        boolean isAuthor = comment.getUser().equals(user);
+        if (isModerator || isAuthor) {
+            commentRepository.deleteById(id);
+            return HttpStatus.NO_CONTENT;
+        }
+        return HttpStatus.FORBIDDEN;
     }
 
     public Page<Comment> getAllCommentsByUserId(Integer userId, Pageable pageable) {
